# go/token: match the behavior of index selection with sort.Search