@shmsr 2021-08-20
Now that's Go 1.17 is released, this could be merged.